chore: fix test cases
